update infra folder structure
