+++ src/modules/user/src/infrastructure/http/controllers/user/updateUserPermissionController.ts
