+++ src/shared/infrastructure/trpc/routes.ts
@@ -1,5 +1,5 @@
 import { createCallerFactory, router } from ".";
-import { userRouter } from "@/modules/user/src/infrastructure/http/routes/userRouter";
+import { userRouter } from "@/modules/user/src/infrastructure/http/routes/user/userRouter";
 
 export const appRouter = router({
   userRouter,
